Consolidate grapher code into one grapher per metric
In the future, different graphing engines will be specified as an
adapter to the metric grapher
