--- lib/metric_fu/metrics/cane/cane_bluff_grapher.rb
@@ -1,16 +1,0 @@
-MetricFu.metrics_require   { 'cane/cane_grapher' }
-module MetricFu
-  class CaneBluffGrapher < CaneGrapher
-    def title
-      'Cane: code quality threshold violations'
-    end
-    def data
-      [
-        ['cane', @cane_violations.join(',')]
-      ]
-    end
-    def output_filename
-      'cane.js'
-    end
-  end
-end
