--- lib/metric_fu/metrics/flay/flay_bluff_grapher.rb
@@ -1,16 +1,0 @@
-MetricFu.metrics_require   { 'flay/flay_grapher' }
-module MetricFu
-  class FlayBluffGrapher < FlayGrapher
-    def title
-      'Flay: duplication'
-    end
-    def data
-      [
-        ['flay', @flay_score.join(',')]
-      ]
-    end
-    def output_filename
-      'flay.js'
-    end
-  end
-end
