+++ lib/metric_fu/metrics/rcov/rcov_grapher.rb
@@ -3,6 +3,10 @@ module MetricFu
   class RcovGrapher < Grapher
     attr_accessor :rcov_percent, :labels
 
+    def self.metric
+      :rcov
+    end
+
     def initialize
       super
       self.rcov_percent = []
@@ -15,5 +19,20 @@ module MetricFu
         self.labels.update( { self.labels.size => date })
       end
     end
+
+    def title
+      'Rcov: code coverage'
+    end
+
+    def data
+      [
+        ['rcov', @rcov_percent.join(',')]
+      ]
+    end
+
+    def output_filename
+      'rcov.js'
+    end
+
   end
 end
