+++ lib/metric_fu/metrics/stats/stats_grapher.rb
@@ -3,6 +3,10 @@ module MetricFu
   class StatsGrapher < Grapher
     attr_accessor :loc_counts, :lot_counts, :labels
 
+    def self.metric
+      :stats
+    end
+
     def initialize
       super
       self.loc_counts = []
@@ -17,5 +21,21 @@ module MetricFu
         self.labels.update( { self.labels.size => date })
       end
     end
+
+    def title
+      'Stats: LOC & LOT'
+    end
+
+    def data
+      [
+        ['LOC', @loc_counts.join(',')],
+        ['LOT', @lot_counts.join(',')],
+      ]
+    end
+
+    def output_filename
+      'stats.js'
+    end
+
   end
 end
